% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromage-package.R
\docType{package}
\name{fromage-package}
\alias{fromage}
\alias{fromage-package}
\title{fromage: The multimodel Q control chart}
\description{
Implements the multimodel Q (\code{QMM}) methodology for monitoring changes in
the serial-dependence structure of an autocorrelated process. Start with
\code{\link[=qmm_chart]{qmm_chart()}} for a complete single-series workflow.
}
\author{
\strong{Maintainer}: A B \email{a.b@example.com}

}
\keyword{internal}
