% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmm-monitor.R
\name{signals}
\alias{signals}
\alias{signals.qmm_monitor}
\title{Extract signalled observations}
\usage{
signals(object, ...)

\method{signals}{qmm_monitor}(object, ...)
}
\arguments{
\item{object}{A monitoring result.}

\item{...}{Additional arguments passed to an object-specific method.}
}
\value{
A data frame containing only signalled horizons.
}
\description{
Extract signalled observations
}
