% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_qmm_calibration}
\alias{validate_qmm_calibration}
\title{Validate a QMM calibration object}
\usage{
validate_qmm_calibration(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x} invisibly. An error is raised if \code{x} does not inherit from
\code{qmm_calibration}.
}
\description{
Checks the S3 class required by functions that consume \code{\link[=calibrate_qmm]{calibrate_qmm()}}
output.
}
\keyword{internal}
