% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_qmm_fit}
\alias{validate_qmm_fit}
\title{Validate a QMM Phase I fit object}
\usage{
validate_qmm_fit(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x} invisibly. An error is raised if \code{x} does not inherit from
\code{qmm_fit}.
}
\description{
Checks the S3 class required by functions that consume \code{\link[=fit_qmm]{fit_qmm()}} output.
}
\keyword{internal}
